PM-28056: Consolidate IntentManager extensions
