+++ app/src/test/kotlin/com/x8bit/bitwarden/ui/vault/feature/manualcodeentry/ManualCodeEntryScreenTests.kt
@@ -1,8 +1,5 @@
 package com.x8bit.bitwarden.ui.vault.feature.manualcodeentry
 
-import android.app.Application
-import android.content.Intent
-import android.net.Uri
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.assertIsNotDisplayed
 import androidx.compose.ui.test.assertTextEquals
@@ -14,9 +11,9 @@ import androidx.compose.ui.test.onFirst
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTextInput
-import androidx.test.core.app.ApplicationProvider
 import com.bitwarden.core.data.repository.util.bufferedMutableSharedFlow
 import com.bitwarden.ui.platform.manager.IntentManager
+import com.bitwarden.ui.platform.manager.util.startAppSettingsActivity
 import com.bitwarden.ui.platform.resource.BitwardenString
 import com.bitwarden.ui.util.asText
 import com.bitwarden.ui.util.assertNoDialogExists
@@ -25,9 +22,8 @@ import com.x8bit.bitwarden.ui.platform.base.BitwardenComposeTest
 import com.x8bit.bitwarden.ui.platform.manager.permissions.FakePermissionManager
 import io.mockk.every
 import io.mockk.mockk
-import io.mockk.slot
+import io.mockk.mockkStatic
 import io.mockk.verify
-import junit.framework.TestCase.assertEquals
 import junit.framework.TestCase.assertTrue
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.update
@@ -78,22 +74,11 @@ class ManualCodeEntryScreenTests : BitwardenComposeTest() {
 
     @Test
     fun `on NavigateToAppSettings event should invoke intent handler`() {
-        mutableEventFlow.tryEmit(ManualCodeEntryEvent.NavigateToAppSettings)
-
-        val uri = Uri.parse(
-            "package:" +
-                ApplicationProvider
-                    .getApplicationContext<Application>()
-                    .packageName,
-        )
-
-        val intentSlot = slot<Intent>()
-        verify { intentManager.startActivity(capture(intentSlot)) }
-
-        assertEquals(
-            uri,
-            intentSlot.captured.data,
-        )
+        mockkStatic(IntentManager::startAppSettingsActivity) {
+            every { intentManager.startAppSettingsActivity() } returns true
+            mutableEventFlow.tryEmit(ManualCodeEntryEvent.NavigateToAppSettings)
+            verify(exactly = 1) { intentManager.startAppSettingsActivity() }
+        }
     }
 
     @Suppress("MaxLineLength")
