+++ authenticator/src/main/kotlin/com/bitwarden/authenticator/ui/platform/util/IntentManagerExtensions.kt
@@ -3,20 +3,9 @@
 package com.bitwarden.authenticator.ui.platform.util
 
 import android.content.Intent
-import android.provider.Settings
 import androidx.core.net.toUri
 import com.bitwarden.annotation.OmitFromCoverage
 import com.bitwarden.ui.platform.manager.IntentManager
-
-/**
- * Launches the authenticator app settings.
- */
-fun IntentManager.startAuthenticatorAppSettings() {
-    startActivity(
-        Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS)
-            .setData("package:$packageName".toUri()),
-    )
-}
 
 /**
  * Launches the Bitwarden account settings.
